moved model and dto code to database folder
